Client To Server Data
We can use this structure of "Wait Target Data" to access client only data such as cursor location and then execute server logic with that uploaded data as seen in Blood Orb here. This also comes with net sync logic so even on 200 ping both players are seeing the exact same thing.